Create drawing pad with 16x16 set of divs
I'm looking to try and make a grid in css with the divs. Each column of
divs is nested in a row of divs. I'm not sure if this is the right way
to do this, but it's my first instinct. Perhaps my instinct is leading
me towards how I could accomplish the same thing with flex.